tests/wpt/web-platform-tests/css/css3-selectors/i18n/css3-selectors-lang-056.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>[lang="es"], xml:lang="es" (html)</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/css3-selectors/#attribute-selectors'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name='flags' content='HTMLonly'>
<style type='text/css'>
.test div { width: 50px; }
#box[lang='es'] { width: 100px; }
</style>
</head>
<body>



<div class="test"><div id="box" xml:lang="es">&#xA0;</div></div>


<script>
test(function() {
assert_equals(document.getElementById('box').offsetWidth, 50);
}, "A [lang='es'] value that matches an identical xml:lang attribute value will NOT produce styling in pages served as HTML.");
</script>

<div id='log'></div>

</body>
</html>
